Move Sb link into PageTemplate
